requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
djangorestframework==3.16.0
injector==0.22.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
